+++ CHANGELOG.md
@@ -78,6 +78,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Separate upload and download workers
   - Use direct file system writes if available on system
   - Move upload, download fetch calls to workers to reduce messaging overhead
+- (GL #1145) Added new api for modifying write access from container's access control list
 
 ### Changed